% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_diverse_explicit}
\alias{fitness_diverse_explicit}
\title{Explicit order fitness function.}
\usage{
fitness_diverse_explicit(x, ranking, maximize = FALSE)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Named vector of at least two performance values.}

\item{ranking}{[\code{character}]\cr
Desired, i.e. explicit, ranking.}

\item{maximize}{[\code{logical(1)}]\cr
Is the goal to maximize performance values?
Defaults to \code{FALSE}.}
}
\value{
[\code{numeric(3)}]
}
\description{
To be written ...
}
