% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_diverse_noorder}
\alias{fitness_diverse_noorder}
\title{No-order fitness function.}
\usage{
fitness_diverse_noorder(x, maximize = FALSE)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Vector of at least two performance values.}

\item{maximize}{[\code{logical(1)}]\cr
Is the goal to maximize performance values?
Defaults to \code{FALSE}.}
}
\value{
[\code{numeric(1)}]
}
\description{
Given performance values \eqn{p_1, \ldots, p_n} of \eqn{n \geq 3}
algorithms, this function first sorts the values in increasing order resulting
in the order statistics \eqn{p_{(1)}, \ldots, p_{(n)}}, i.e. \eqn{p_{(i)}} is
the \eqn{i}th largest value. Next the function calculates the scalar fitness
value implementing the following formula:
\eqn{
f(p_1, \ldots, p_n) = \sum_{i=2}^{n-1} \left(p_{(i+1)} - p_{(i)}\right) \cdot \left(p_{(i)} - p_{(i-1)}\right).
}
}
