% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{get_algorithm_ranking}
\alias{get_algorithm_ranking}
\title{Get algorithm ranking.}
\usage{
get_algorithm_ranking(x, maximize = FALSE, as.string = FALSE, sep = "---")
}
\arguments{
\item{x}{[\code{numeric} | \code{list of numeric vectors}]\cr
Named vector of at least two performance values or list of such.}

\item{maximize}{[\code{logical(1)}]\cr
Is the goal to maximize performance values?
Defaults to \code{FALSE}.}

\item{as.string}{[\code{logical(1L)}]\cr
Convert to string representation, i.e. \code{c("A", "B", "C")} to
\code{"A---B---C"}.}

\item{sep}{[\code{character(1)}]\cr
Separator used to \dQuote{glue together} if \code{as.string = TRUE}.
Default is \dQuote{---}.}
}
\value{
[\code{character}] \code{names(x)} in order of performance.
}
\description{
Given a named vector of performance values the function returns
the names in sorted order of performance.
}
