% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{get_algorithm_ranking}
\alias{get_algorithm_ranking}
\title{Get algorithm ranking.}
\usage{
get_algorithm_ranking(x, maximize = FALSE)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Vector of at least two performance values.}

\item{maximize}{[\code{logical(1)}]\cr
Is the goal to maximize performance values?
Defaults to \code{FALSE}.}
}
\value{
[\code{character}] \code{names(x)} in order of performance.
}
\description{
Given a named vector of performance values the function returns
the names in sorted order of performance.
}
